Update run script. Filename input.

+++ run.sh
@@ -1,3 +1,9 @@
+if [ $# -eq 0 ]
+then
+    echo "Filename not supplied"
+    exit 1
+fi
+
 sudo rm iperf*
 sudo mn -c
-sudo python build_topology.py
+sudo python $1
